feat: make improvements
