+++ .gitignore
@@ -30,6 +30,7 @@
 .pub/
 build/
 coverage/
+pubspec.lock
 
 # Android related
 **/android/**/gradle-wrapper.jar
