+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## [0.0.1] - TODO: Add release date.
+## 0.1.0
+
+- Upgrade dependencies and improve API
+
+## 0.0.1
 
 - Initial Release!
