+++ example/pubspec.yaml
@@ -3,7 +3,7 @@ description: An example of using Validate Firebase Auth in Dart Functions Framew
 publish_to: none
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   functions_framework: ^0.4.0
@@ -11,8 +11,9 @@ dependencies:
   validate_firebase_auth: ^0.0.1
 
 dev_dependencies:
-  build_runner: ^1.10.7
+  build_runner: ^2.4.7
   functions_framework_builder: ^0.4.0
   http: ^0.13.0
   test: ^1.15.7
   test_process: ^2.0.0
+  very_good_analysis: ^2.0.0
