+++ lib/validate_firebase_auth.dart
@@ -1,4 +1,7 @@
+/// Validates Firebase auth tokens for verifying authentication on server-side
+/// applications.
 library validate_firebase_auth;
 
 export 'package:openid_client/openid_client.dart' show IdToken;
+
 export 'src/validate_firebase_auth.dart' show FirebaseAuthValidator;
