+++ pubspec.yaml
@@ -1,18 +1,18 @@
 name: validate_firebase_auth
 description: Validates Firebase auth tokens for verifying authentication on server-side applications.
-version: 0.0.1
+version: 0.1.0
 repository: https://github.com/mtwichel/validate-firebase-auth-dart
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
-  very_good_analysis: ^2.0.0
+  http: ^1.0.0
+  meta: ^1.0.0
   openid_client: ^0.4.3
-  http: ^0.13.3
-  meta: ^1.4.0
 
 dev_dependencies:
-  mocktail: ^0.1.1
-  test: ^1.16.8
   coverage: ^1.0.2
+  mocktail: ^1.0.2
+  test: ^1.16.8
+  very_good_analysis: ^5.1.0
